Add check if firebase reference is valid.

+++ mapswipe_workers/tests/unittests/tear_down.py
@@ -1,4 +1,5 @@
 """Helper functions for test tear down"""
+import re
 
 from mapswipe_workers import auth
 
@@ -8,6 +9,12 @@ def delete_test_data(project_id: str) -> None:
     Delete test project indluding groups, tasks and results
     from Firebase and Postgres
     """
+
+    if not re.match(r"[-a-zA-Z0-9]+", project_id):
+        raise ValueError(
+            f"Given argument resulted in invalid Firebase Realtime Database reference. "
+        )
+
     fb_db = auth.firebaseDB()
     ref = fb_db.reference("v2/results/{0}".format(project_id))
     ref.delete()
